+++ CHANGES.md
@@ -1,6 +1,10 @@
 # CHANGES
 ## RosarioSIS Student Information System
 
+Changes in 4.9.4
+----------------
+- Fix regression since v4.3 translated Help file include in Help.fnc.php
+
 Changes in 4.9.3
 ----------------
 - Fix regression since v3.4.2 "Fix SQL error if MP was deleted" in ScheduleReport.php